requirements.txt
# Core Flask dependencies
flask==3.1.2
werkzeug==3.1.3

# Machine Learning and Computer Vision
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.0.0
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2

# Cloud API integration (optional)
google-cloud-vision==3.4.4

# Data processing and utilities
pandas==2.3.3
scipy==1.16.2

# Development and testing
pytest>=8.0.0
pytest-flask>=1.3.0
